+++ joblib/externals/loky/process_executor.py
@@ -4,7 +4,6 @@
 # author: Thomas Moreau and Olivier Grisel
 #
 # adapted from concurrent/futures/process_pool_executor.py (17/02/2017)
-#  * Backport for python2.7/3.3,
 #  * Add an extra management thread to detect executor_manager_thread failures,
 #  * Improve the shutdown process to avoid deadlocks,
 #  * Add timeout for workers,
@@ -62,6 +61,7 @@ __author__ = 'Thomas Moreau (thomas.moreau.2010@gmail.com)'
 import os
 import gc
 import sys
+import queue
 import struct
 import weakref
 import warnings
@@ -72,27 +72,18 @@ from time import time
 import multiprocessing as mp
 from functools import partial
 from pickle import PicklingError
+from concurrent.futures import Executor
+from concurrent.futures._base import LOGGER
+from concurrent.futures.process import BrokenProcessPool as _BPPException
+from multiprocessing.connection import wait
 
-from . import _base
+from ._base import Future
 from .backend import get_context
-from .backend.compat import queue
-from .backend.compat import wait
-from .backend.compat import set_cause
 from .backend.context import cpu_count
 from .backend.queues import Queue, SimpleQueue
 from .backend.reduction import set_loky_pickler, get_loky_pickler_name
-from .backend.utils import recursive_terminate, get_exitcodes_terminated_worker
+from .backend.utils import kill_process_tree, get_exitcodes_terminated_worker
 from .initializers import _prepare_initializer
-
-try:
-    from concurrent.futures.process import BrokenProcessPool as _BPPException
-except ImportError:
-    _BPPException = RuntimeError
-
-
-# Compatibility for python2.7
-if sys.version_info[0] == 2:
-    ProcessLookupError = OSError
 
 
 # Mechanism to prevent infinite process spawning. When a worker of a
@@ -117,7 +108,7 @@ try:
             gc.collect()
 
         mem_size = Process(pid).memory_info().rss
-        mp.util.debug('psutil return memory size: {}'.format(mem_size))
+        mp.util.debug(f'psutil return memory size: {mem_size}')
         return mem_size
 
 except ImportError:
@@ -137,12 +128,7 @@ class _ThreadWakeup:
 
     def wakeup(self):
         if not self._closed:
-            if sys.platform == "win32" and sys.version_info[:2] < (3, 4):
-                # Compat for python2.7 on windows, where poll return false for
-                # b"" messages. Use the slightly larger message b"0".
-                self._writer.send_bytes(b"0")
-            else:
-                self._writer.send_bytes(b"")
+            self._writer.send_bytes(b"")
 
     def clear(self):
         if not self._closed:
@@ -150,7 +136,7 @@ class _ThreadWakeup:
                 self._reader.recv_bytes()
 
 
-class _ExecutorFlags(object):
+class _ExecutorFlags:
     """necessary references to maintain executor states without preventing gc
 
     It permits to keep the information needed by executor_manager_thread
@@ -199,7 +185,7 @@ def _python_exit():
     _global_shutdown = True
     items = list(_threads_wakeups.items())
     mp.util.debug("Interpreter shutting down. Waking up "
-                  "executor_manager_thread {}".format(items))
+                  f"executor_manager_thread {items}")
     for _, (shutdown_lock, thread_wakeup) in items:
         with shutdown_lock:
             thread_wakeup.wakeup()
@@ -227,7 +213,7 @@ class _RemoteTraceback(Exception):
     """Embed stringification of remote traceback in local traceback
     """
     def __init__(self, tb=None):
-        self.tb = '\n"""\n{}"""'.format(tb)
+        self.tb = f'\n"""\n{tb}"""'
 
     def __str__(self):
         return self.tb
@@ -249,11 +235,11 @@ class _ExceptionWithTraceback(BaseException):
 
 
 def _rebuild_exc(exc, tb):
-    exc = set_cause(exc, _RemoteTraceback(tb))
+    exc.__cause__ = _RemoteTraceback(tb)
     return exc
 
 
-class _WorkItem(object):
+class _WorkItem:
 
     __slots__ = ["future", "fn", "args", "kwargs"]
 
@@ -264,7 +250,7 @@ class _WorkItem(object):
         self.kwargs = kwargs
 
 
-class _ResultItem(object):
+class _ResultItem:
 
     def __init__(self, work_id, exception=None, result=None):
         self.work_id = work_id
@@ -272,7 +258,7 @@ class _ResultItem(object):
         self.result = result
 
 
-class _CallItem(object):
+class _CallItem:
 
     def __init__(self, work_id, fn, args, kwargs):
         self.work_id = work_id
@@ -288,8 +274,9 @@ class _CallItem(object):
         return self.fn(*self.args, **self.kwargs)
 
     def __repr__(self):
-        return "CallItem({}, {}, {}, {})".format(
-            self.work_id, self.fn, self.args, self.kwargs)
+        return (
+            f"CallItem({self.work_id}, {self.fn}, {self.args}, {self.kwargs})"
+        )
 
 
 class _SafeQueue(Queue):
@@ -299,7 +286,7 @@ class _SafeQueue(Queue):
         self.thread_wakeup = thread_wakeup
         self.pending_work_items = pending_work_items
         self.running_work_items = running_work_items
-        super(_SafeQueue, self).__init__(max_size, reducers=reducers, ctx=ctx)
+        super().__init__(max_size, reducers=reducers, ctx=ctx)
 
     def _on_queue_feeder_error(self, e, obj):
         if isinstance(obj, _CallItem):
@@ -313,8 +300,7 @@ class _SafeQueue(Queue):
                     "Could not pickle the task to send it to the workers.")
             tb = traceback.format_exception(
                 type(e), e, getattr(e, "__traceback__", None))
-            raised_error = set_cause(raised_error,
-                                     _RemoteTraceback(''.join(tb)))
+            raised_error.__cause__ = _RemoteTraceback(''.join(tb))
             work_item = self.pending_work_items.pop(obj.work_id, None)
             self.running_work_items.remove(obj.work_id)
             # work_item can be None if another process terminated. In this
@@ -325,15 +311,12 @@ class _SafeQueue(Queue):
                 del work_item
             self.thread_wakeup.wakeup()
         else:
-            super(_SafeQueue, self)._on_queue_feeder_error(e, obj)
+            super()._on_queue_feeder_error(e, obj)
 
 
 def _get_chunks(chunksize, *iterables):
     """Iterates over zip()ed iterables in chunks. """
-    if sys.version_info < (3, 3):
-        it = itertools.izip(*iterables)
-    else:
-        it = zip(*iterables)
+    it = zip(*iterables)
     while True:
         chunk = tuple(itertools.islice(it, chunksize))
         if not chunk:
@@ -389,7 +372,7 @@ def _process_worker(call_queue, result_queue, initializer, initargs,
         try:
             initializer(*initargs)
         except BaseException:
-            _base.LOGGER.critical('Exception in initializer:', exc_info=True)
+            LOGGER.critical('Exception in initializer:', exc_info=True)
             # The parent will notice that the process stopped and
             # mark the pool broken
             return
@@ -401,15 +384,14 @@ def _process_worker(call_queue, result_queue, initializer, initargs,
     _last_memory_leak_check = None
     pid = os.getpid()
 
-    mp.util.debug('Worker started with timeout=%s' % timeout)
+    mp.util.debug(f'Worker started with timeout={timeout}')
     while True:
         try:
             call_item = call_queue.get(block=True, timeout=timeout)
             if call_item is None:
                 mp.util.info("Shutting down worker on sentinel")
         except queue.Empty:
-            mp.util.info("Shutting down worker after timeout %0.3fs"
-                         % timeout)
+            mp.util.info(f"Shutting down worker after timeout {timeout:0.3f}s")
             if processes_management_lock.acquire(block=False):
                 processes_management_lock.release()
                 call_item = None
@@ -477,7 +459,7 @@ def _process_worker(call_queue, result_queue, initializer, initargs,
         else:
             # if psutil is not installed, trigger gc.collect events
             # regularly to limit potential memory leaks due to reference cycles
-            if ((_last_memory_leak_check is None) or
+            if (_last_memory_leak_check is None or
                     (time() - _last_memory_leak_check >
                      _MEMORY_LEAK_CHECK_DELAY)):
                 gc.collect()
@@ -514,8 +496,12 @@ class _ExecutorManagerThread(threading.Thread):
         def weakref_cb(_,
                        thread_wakeup=self.thread_wakeup,
                        shutdown_lock=self.shutdown_lock):
-            mp.util.debug('Executor collected: triggering callback for'
-                          ' QueueManager wakeup')
+            if mp is not None:
+                # At this point, the multiprocessing module can already be
+                # garbage collected. We only log debug info when still
+                # possible.
+                mp.util.debug('Executor collected: triggering callback for'
+                              ' QueueManager wakeup')
             with shutdown_lock:
                 thread_wakeup.wakeup()
 
@@ -548,9 +534,7 @@ class _ExecutorManagerThread(threading.Thread):
         # of new processes or shut down
         self.processes_management_lock = executor._processes_management_lock
 
-        super(_ExecutorManagerThread, self).__init__(
-            name="ExecutorManagerThread"
-        )
+        super().__init__(name="ExecutorManagerThread")
         if sys.version_info < (3, 9):
             self.daemon = True
 
@@ -627,7 +611,7 @@ class _ExecutorManagerThread(threading.Thread):
                         "A task has failed to un-serialize. Please ensure that"
                         " the arguments of the function are all picklable."
                     )
-                    set_cause(bpe, result_item)
+                    bpe.__cause__ = result_item
                 else:
                     is_broken = False
             except BaseException as e:
@@ -638,7 +622,7 @@ class _ExecutorManagerThread(threading.Thread):
                 )
                 tb = traceback.format_exception(
                     type(e), e, getattr(e, "__traceback__", None))
-                set_cause(bpe,  _RemoteTraceback(''.join(tb)))
+                bpe.__cause__ = _RemoteTraceback(''.join(tb))
 
         elif wakeup_reader in ready:
             # This is simply a wake-up event that might either trigger putting
@@ -652,8 +636,9 @@ class _ExecutorManagerThread(threading.Thread):
                 # In Windows, introspecting terminated workers exitcodes seems
                 # unstable, therefore they are not appended in the exception
                 # message.
-                exit_codes = "\nThe exit codes of the workers are {}".format(
-                    get_exitcodes_terminated_worker(self.processes)
+                exit_codes = (
+                    "\nThe exit codes of the workers are "
+                    f"{get_exitcodes_terminated_worker(self.processes)}"
                 )
             mp.util.debug('A worker unexpectedly terminated. Workers that '
                           'might have caused the breakage: '
@@ -665,7 +650,7 @@ class _ExecutorManagerThread(threading.Thread):
                 "terminated. This could be caused by a segmentation fault "
                 "while calling the function or by an excessive memory usage "
                 "causing the Operating System to kill the worker.\n"
-                "{}".format(exit_codes)
+                f"{exit_codes}"
             )
 
         self.thread_wakeup.clear()
@@ -739,7 +724,7 @@ class _ExecutorManagerThread(threading.Thread):
         self.executor_flags.flag_as_broken(bpe)
 
         # Mark pending tasks as failed.
-        for work_id, work_item in self.pending_work_items.items():
+        for work_item in self.pending_work_items.values():
             work_item.future.set_exception(bpe)
             # Delete references to object. See issue16284
             del work_item
@@ -775,10 +760,9 @@ class _ExecutorManagerThread(threading.Thread):
         # nested parallelism.
         while self.processes:
             _, p = self.processes.popitem()
-            mp.util.debug("terminate process {}, reason: {}"
-                          .format(p.name, reason))
+            mp.util.debug(f"terminate process {p.name}, reason: {reason}")
             try:
-                recursive_terminate(p)
+                kill_process_tree(p)
             except ProcessLookupError:  # pragma: no cover
                 pass
 
@@ -801,7 +785,7 @@ class _ExecutorManagerThread(threading.Thread):
         n_sentinels_sent = 0
         while (n_sentinels_sent < n_children_to_stop
                 and self.get_n_children_alive() > 0):
-            for i in range(n_children_to_stop - n_sentinels_sent):
+            for _ in range(n_children_to_stop - n_sentinels_sent):
                 try:
                     self.call_queue.put_nowait(None)
                     n_sentinels_sent += 1
@@ -835,7 +819,7 @@ class _ExecutorManagerThread(threading.Thread):
             p.join()
 
         mp.util.debug("executor management thread clean shutdown of worker "
-                      "processes: {}".format(list(self.processes)))
+                      f"processes: {list(self.processes)}")
 
     def get_n_children_alive(self):
         # This is an upper bound on the number of children alive.
@@ -849,9 +833,8 @@ _system_limited = None
 
 def _check_system_limits():
     global _system_limits_checked, _system_limited
-    if _system_limits_checked:
-        if _system_limited:
-            raise NotImplementedError(_system_limited)
+    if _system_limits_checked and _system_limited:
+        raise NotImplementedError(_system_limited)
     _system_limits_checked = True
     try:
         nsems_max = os.sysconf("SC_SEM_NSEMS_MAX")
@@ -866,8 +849,10 @@ def _check_system_limits():
         # minimum number of semaphores available
         # according to POSIX
         return
-    _system_limited = ("system provides too few semaphores (%d available, "
-                       "256 necessary)" % nsems_max)
+    _system_limited = (
+        f"system provides too few semaphores ({nsems_max} available, "
+        "256 necessary)"
+    )
     raise NotImplementedError(_system_limited)
 
 
@@ -895,8 +880,8 @@ def _check_max_depth(context):
     if 0 < MAX_DEPTH and _CURRENT_DEPTH + 1 > MAX_DEPTH:
         raise LokyRecursionError(
             "Could not spawn extra nested processes at depth superior to "
-            "MAX_DEPTH={}. If this is intendend, you can change this limit "
-            "with the LOKY_MAX_DEPTH environment variable.".format(MAX_DEPTH))
+            f"MAX_DEPTH={MAX_DEPTH}. If this is intendend, you can change "
+            "this limit with the LOKY_MAX_DEPTH environment variable.")
 
 
 class LokyRecursionError(RuntimeError):
@@ -933,7 +918,7 @@ class ShutdownExecutorError(RuntimeError):
     """
 
 
-class ProcessPoolExecutor(_base.Executor):
+class ProcessPoolExecutor(Executor):
 
     _at_exit = None
 
@@ -962,8 +947,7 @@ class ProcessPoolExecutor(_base.Executor):
             initargs: A tuple of arguments to pass to the initializer.
             env: A dict of environment variable to overwrite in the child
                 process. The environment variables are set before any module is
-                loaded. Note that this only works with the loky context and it
-                is unreliable under windows with Python < 3.6.
+                loaded. Note that this only works with the loky context.
         """
         _check_system_limits()
 
@@ -1048,17 +1032,6 @@ class ProcessPoolExecutor(_base.Executor):
         if self._executor_manager_thread is None:
             mp.util.debug('_start_executor_manager_thread called')
 
-            # When the executor gets garbage collected, the weakref callback
-            # will wake up the queue management thread so that it can terminate
-            # if there is no pending work item.
-            def weakref_cb(
-                    _, thread_wakeup=self._executor_manager_thread_wakeup,
-                    shutdown_lock=self._shutdown_lock):
-                mp.util.debug('Executor collected: triggering callback for'
-                              ' QueueManager wakeup')
-                with self._shutdown_lock:
-                    thread_wakeup.wakeup()
-
             # Start the processes so that their sentinels are known.
             self._executor_manager_thread = _ExecutorManagerThread(self)
             self._executor_manager_thread.start()
@@ -1099,7 +1072,7 @@ class ProcessPoolExecutor(_base.Executor):
             p._worker_exit_lock = worker_exit_lock
             p.start()
             self._processes[p.pid] = p
-        mp.util.debug('Adjust process count : {}'.format(self._processes))
+        mp.util.debug(f'Adjust process count : {self._processes}')
 
     def _ensure_executor_running(self):
         """ensures all workers and management thread are running
@@ -1123,7 +1096,7 @@ class ProcessPoolExecutor(_base.Executor):
                 raise RuntimeError('cannot schedule new futures after '
                                    'interpreter shutdown')
 
-            f = _base.Future()
+            f = Future()
             w = _WorkItem(f, fn, args, kwargs)
 
             self._pending_work_items[self._queue_count] = w
@@ -1134,7 +1107,7 @@ class ProcessPoolExecutor(_base.Executor):
 
             self._ensure_executor_running()
             return f
-    submit.__doc__ = _base.Executor.submit.__doc__
+    submit.__doc__ = Executor.submit.__doc__
 
     def map(self, fn, *iterables, **kwargs):
         """Returns an iterator equivalent to map(fn, iter).
@@ -1163,14 +1136,14 @@ class ProcessPoolExecutor(_base.Executor):
         if chunksize < 1:
             raise ValueError("chunksize must be >= 1.")
 
-        results = super(ProcessPoolExecutor, self).map(
+        results = super().map(
             partial(_process_chunk, fn), _get_chunks(chunksize, *iterables),
             timeout=timeout
         )
         return _chain_from_iterable_of_lists(results)
 
     def shutdown(self, wait=True, kill_workers=False):
-        mp.util.debug('shutting down executor %s' % self)
+        mp.util.debug(f'shutting down executor {self}')
 
         self._flags.flag_as_shutting_down(kill_workers)
         executor_manager_thread = self._executor_manager_thread
@@ -1192,4 +1165,4 @@ class ProcessPoolExecutor(_base.Executor):
         self._result_queue = None
         self._processes_management_lock = None
 
-    shutdown.__doc__ = _base.Executor.shutdown.__doc__
+    shutdown.__doc__ = Executor.shutdown.__doc__